support go1.9
