+++ v2/internal/errors/errors_1_9.go
@@ -1,0 +1,16 @@
+// +build !go1.10
+
+package errors
+
+import "bytes"
+
+func (e multierr) Error() string {
+	var b bytes.Buffer
+	for i, err := range e {
+		if i != 0 {
+			b.WriteString("; ")
+		}
+		b.WriteString(err.Error())
+	}
+	return b.String()
+}
